feat(CRUD): change access to specific users

+++ backend/controllers/pantry.controller.js
@@ -1,8 +1,14 @@
 const pool = require("../db");
 
+// User object is passed in through req from authentication step
+
 exports.getAllItems = async (req, res) => {
   try {
-    const result = await pool.query("SELECT * FROM pantry_items");
+    const { user } = req;
+    const result = await pool.query(
+      "SELECT * FROM pantry_items WHERE user_id = $1",
+      [user.userId],
+    );
     res.json(result.rows);
   } catch (err) {
     console.error(err);
@@ -12,10 +18,11 @@ exports.getAllItems = async (req, res) => {
 
 exports.addItem = async (req, res) => {
   try {
+    const { user } = req;
     const { name, quantity, unit } = req.body;
     const result = await pool.query(
-      "INSERT INTO pantry_items (name, quantity, unit) VALUES ($1, $2, $3) RETURNING *",
-      [name, quantity, unit],
+      "INSERT INTO pantry_items (name, quantity, unit, user_id) VALUES ($1, $2, $3, $4) RETURNING *",
+      [name, quantity, unit, user.userId],
     );
     res.status(201).json(result.rows[0]);
   } catch (err) {
@@ -25,27 +32,29 @@ exports.addItem = async (req, res) => {
 
 exports.updateItem = async (req, res) => {
   try {
+    const { user } = req;
     const { id } = req.params;
     const { name, quantity, unit } = req.body;
     const result = await pool.query(
-      "UPDATE pantry_items SET name = $1, quantity = $2, unit = $3 WHERE id = $4 RETURNING *",
-      [name, quantity, unit, id],
+      "UPDATE pantry_items SET name = $1, quantity = $2, unit = $3 WHERE id = $4 AND user_id = $5 RETURNING *",
+      [name, quantity, unit, id, user.userId],
     );
     if (result.rows.length === 0) {
       return res.status(404).json({ error: "Item not found" });
     }
     res.json(result.rows[0]);
   } catch (err) {
-    res.status(500).json({ error: err });
+    res.status(500).json({ error: err.message });
   }
 };
 
 exports.deleteItem = async (req, res) => {
   try {
+    const { user } = req;
     const { id } = req.params;
     const result = await pool.query(
-      "DELETE FROM pantry_items WHERE id = $1 RETURNING *",
-      [id],
+      "DELETE FROM pantry_items WHERE id = $1 AND user_id = $2 RETURNING *",
+      [id, user.userId],
     );
     if (result.rows.length === 0) {
       return res.status(404).json({ error: "Item not found" });
